Adding Connect content
