+++ client/src/pages/components/Header/Header.module.css
@@ -13,7 +13,7 @@
     color: rgb(var(--space-white-rgb));
     vertical-align: center;
     font-size: 3rem;
-    font-family: monospace;
+    font-family: 'Roboto Mono', monospace;
     font-weight: normal;
     padding-left: 1rem;
     white-space: nowrap;
