+++ client/src/pages/components/MainContainer/MainContainer.module.css
@@ -1,5 +1,5 @@
 .mainContainer {
-    max-width: 1200px;
+    max-width: 1000px;
     width: 100%;
     background-image: radial-gradient(ellipse, rgba(var(--space-black-rgb), 0), rgba(var(--space-black-rgb), 1)),
     url("../../../imgs/cosmic-cliffs.png");
