+++ client/src/pages/components/MainContent/Connect.tsx
@@ -1,0 +1,12 @@
+import ContentHeader from "@/pages/components/MainContent/ContentCard/ContentHeader";
+import {TbPlugConnected} from "react-icons/tb";
+
+export default function Connect(){
+
+    return(
+        <div id={"connect"}>
+            <ContentHeader title={"Connect"} icon={<TbPlugConnected/>}/>
+
+        </div>
+    )
+}
