+++ client/src/pages/hooks/useNavBar.tsx
